refactor: 메인 페이지에 리팩토링된 예약일 지정 UI 연결
