+++ src/components/datepicker/Weekdays.tsx
@@ -4,24 +4,26 @@ import styled from 'styled-components';
 
 function Weekdays() {
   return (
-    <Container>
-      {DISPLAY_WEEKDAYS.map((day) => {
-        return <Weekday key={day}>{day}</Weekday>;
+    <WeekdayRow>
+      {DISPLAY_WEEKDAYS.map((weekday) => {
+        return <Weekday key={weekday}>{weekday}</Weekday>;
       })}
-    </Container>
+    </WeekdayRow>
   );
 }
 
 export default Weekdays;
 
-const Container = styled.div`
-  margin-top: 8px;
+const WeekdayRow = styled.div`
+  width: 100%;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
 `;
+
 const Weekday = styled.div`
-  flex: 1;
+  width: calc(100% / 7);
   text-align: center;
-  text-align: center;
+  margin-top: 8px;
+  margin-bottom: 8px;
 `;
